feat(agent): Introduce dedicated message handling for agent runs
- Separates the message flow and UI components between standard chat and agentic runs.
- Introduces AgentTurnMessageDTO for agent-specific messaging data.
- Updates AgentRunViewModel to use the new agent-specific DTOs.
- Refactors chat views to utilize a distinct chat message list component.
